This version should allow you to visualize marker array, load customers from yaml, and drive to the customers in the yaml sequence, then head home when done (0, 0, 0).
Had to move the MarkerArray to the customer_generator as I think I was running into timing issues
that ultimately were killing way_point_manager because everytime I tried to access the x,y from
the PoseArray published from customer_generator to way_point_manager that is when the way_point_manager would crash.
Anyway its probably better that customer_generator pushlishes the visualizaiton since it is responisble for the customer's
pose anyway for navigation.